large-text-summarizer/requirements.txt
# Large Text Summarizer - Dependencies

# =============================================================================
# LOCAL LLM Integration
# =============================================================================
langchain>=0.1.0
langchain-ollama>=0.0.1
langchain-core>=0.1.0

# =============================================================================
# Token Counting
# =============================================================================
tiktoken>=0.5.0

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# MCP Server (Optional - for MCP integration)
# =============================================================================
mcp>=0.1.0

# =============================================================================
# Type Hints
# =============================================================================
typing-extensions>=4.0.0
